requirements.txt
Flask
tensorflow-macos
tensorflow-metal
opencv-python
matplotlib
numpy